refactor split soak testing code
